Move example files to an examples package
