+++ c3r-sdk-examples/src/main/java/com/amazonaws/c3r/examples/parquet/package-info.java
@@ -8,4 +8,4 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.amazonaws.c3r.examples;
+package com.amazonaws.c3r.examples.parquet;
