+++ c3r-sdk-examples/src/test/java/com/amazonaws/c3r/examples/parquet/ParquetExampleTest.java
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.amazonaws.c3r.examples;
+package com.amazonaws.c3r.examples.parquet;
 
 import org.junit.jupiter.api.Test;
 
@@ -19,11 +19,11 @@ public class ParquetExampleTest {
         final Path encryptedParquet = Files.createTempFile("encrypted", ".parquet");
         final Path decryptedParquet = Files.createTempFile("decrypted", ".parquet");
 
-        com.amazonaws.c3r.examples.ParquetExample.encrypt(inputParquet.toString(), encryptedParquet.toString());
+        ParquetExample.encrypt(inputParquet.toString(), encryptedParquet.toString());
         assertTrue(Files.exists(encryptedParquet));
         assertTrue(Files.size(encryptedParquet) > 0);
 
-        com.amazonaws.c3r.examples.ParquetExample.decrypt(encryptedParquet.toString(), decryptedParquet.toString());
+        ParquetExample.decrypt(encryptedParquet.toString(), decryptedParquet.toString());
 
         assertTrue(Files.exists(decryptedParquet));
         assertTrue(Files.size(decryptedParquet) > 0);
