+++ c3r-sdk-parquet/build.gradle
@@ -36,29 +36,12 @@ plugins {
 
 apply plugin: 'com.github.spotbugs'
 
-sourceSets {
-    examples {
-        compileClasspath += sourceSets.main.output
-        runtimeClasspath += sourceSets.main.output
-        java {
-            srcDirs = ['src/examples']
-        }
-    }
-    // test needs to access classes in the examples directory so adding a dependency
-    test {
-        compileClasspath += sourceSets.main.output
-        compileClasspath += sourceSets.examples.output
-        runtimeClasspath += sourceSets.main.output
-        runtimeClasspath += sourceSets.examples.output
-    }
-}
-
 /*
  Configures the Checkstyle "checkstyle" plugin. Remove this and the plugin if
  you want to skip these checks and report generation.
 */
 checkstyle {
-    sourceSets = [sourceSets.main, sourceSets.examples, sourceSets.test]
+    sourceSets = [sourceSets.main, sourceSets.test]
     configFile = file('../config/checkstyle/checkstyle.xml')
     configProperties.put('checkstyle.suppression.filter', '../config/checkstyle/suppressions.xml')
     configDirectory.set(file('../config/checkstyle'))
@@ -108,12 +91,6 @@ dependencies {
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.2'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.9.2'
     testImplementation project(path: ':c3r-sdk-core', configuration: 'testArtifacts')
-}
-
-configurations {
-    examples {
-        examplesImplementation.extendsFrom implementation
-    }
 }
 
 test {
