+++ settings.gradle
@@ -13,3 +13,4 @@ rootProject.name = 'cryptographic-computing-for-clean-rooms'
 include 'c3r-sdk-core'
 include 'c3r-sdk-parquet'
 include 'c3r-cli'
+include 'c3r-sdk-examples'
